+++ Test/Input/BobrilSourceInfo/StyleDef/out/index.js
@@ -1,0 +1,8 @@
+"use strict";
+var _a;
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.s = void 0;
+var b = require("bobril");
+var k = "k";
+exports.s = b.styleDef({ color: "blue" }, (_a = {}, _a[".".concat(k)] = { color: "red" }, _a), "name");
+//# sourceMappingURL=index.js.map