+++ Test/Input/Bundler/LibrarySimple/index.js
@@ -1,0 +1,7 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.getDescriptor = void 0;
+function getDescriptor() {
+    return { asd: "df" };
+}
+exports.getDescriptor = getDescriptor;