+++ Test/Input/Bundler/LibrarySimple/index.ts
@@ -1,0 +1,3 @@
+export function getDescriptor() {
+    return { asd: "df" };
+}