+++ Test/Input/Bundler/LibrarySimple/out/cb-bundle.js
@@ -1,0 +1,8 @@
+function getDescriptor() {
+    return {
+        asd: "df"
+    };
+}
+
+export { getDescriptor };
+