+++ Test/Input/Bundler/LibrarySimple/out/cbm-bundle.js
@@ -1,0 +1,8 @@
+function t() {
+    return {
+        asd: "df"
+    };
+}
+
+export { t as getDescriptor };
+