+++ Test/Input/Bundler/LibrarySimple/out/cm-bundle.js
@@ -1,0 +1,1 @@
+function t(){return{asd:"df"}}export{t as getDescriptor}