+++ Test/Input/Bundler/LibraryStaticImport/index.js
@@ -1,0 +1,10 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.addOne = void 0;
+var lib2_1 = require("./lib2");
+function addOne(v) {
+    return (0, lib2_1.bar)(1, v);
+}
+exports.addOne = addOne;
+console.log((0, lib2_1.bar)(1, 2));
+exports.default = lib2_1.bar;