+++ Test/Input/Bundler/LibraryStaticImport/lib2.js
@@ -1,0 +1,5 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.bar = void 0;
+var lib_1 = require("./lib");
+Object.defineProperty(exports, "bar", { enumerable: true, get: function () { return lib_1.fn; } });